Add files via upload
This is the json file used to perform Data Cleaning aThis repository contains both the json and csv files. As my Mac supports JSON but not CSV to upload on the SQL server, I have used JSON files for data cleaning and EDA. You can find the CSV file in this repository as well!